pyproject.toml
[project]
name = "pysswordsz"
version = "0.1.0"
description = "A password manager in Python that not only manages passwords but also serves as a simple encryption software."
authors = [
    { name = "SidneyZhang", email = "zly@lyzhang.me" }
]
dependencies = [
    "typer>=0.12.3",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pysswordsz"]
